requirements.txt
loguru==0.6.0
peewee==3.14.10
pyTelegramBotAPI==4.4.0
requests==2.27.1
python-dotenv==0.19.2
